Store unknown Attributes in a generic container
I want to store all attributes I receive from an API into a database, even if they are not properties on my C#/CLR classes. 
I tried using JsonExtensionData on one of my models, but this did not work.

So I decided to put a extraAttributes dict on my base ResourceObject class which is now filled up with non-mapped properties which I can then store in my database.